detection_server/requirements.txt
Flask
gunicorn
google-cloud
google-cloud-storage
Pillow
tensorflow==2.9.2
numpy
wheel
waitress